+++ modules/access/http.c
@@ -1074,7 +1074,7 @@ static int Control( access_t *p_access, int i_query, va_list args )
         /* */
         case ACCESS_GET_PTS_DELAY:
             pi_64 = (int64_t*)va_arg( args, int64_t * );
-            *pi_64 = (int64_t)var_GetInteger( p_access, "http-caching" ) * 1000;
+            *pi_64 = var_GetInteger( p_access, "http-caching" ) * 1000;
             break;
 
         /* */
