+++ modules/access/smb.c
@@ -335,7 +335,7 @@ static int Control( access_t *p_access, int i_query, va_list args )
 
     case ACCESS_GET_PTS_DELAY:
         *va_arg( args, int64_t * )
-                  = (int64_t)var_GetInteger( p_access, "smb-caching" ) * 1000;
+                  = var_GetInteger( p_access, "smb-caching" ) * 1000;
         break;
 
     case ACCESS_SET_PAUSE_STATE:
